+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+In Development
+--------------
+- Update minimum pydantic version to 1.9
+
 v0.1.1 (2022-01-02)
 -------------------
 - Support tomli 2.0